Some more refactoring
Abstracted the data layer into a repo to help separate the data layer as well as make it possible to mock and test.

Added dependency injection for ReferralsService.cs to allow for mocking and testing however left a basic constructor as I'm restricted from changing Program.cs.

Moved "magic strings/ints" to help avoid problems with them in the future.

Created better tests and mocked the data layer for tests.

Added InternalsVisibileTo the assembly for the testing to allow me to handle internal varibales, methods and classes from the test project.